Working DB pulls for base homepage, fleshing out settings next

+++ flask/bin/downloader.py
@@ -8,10 +8,12 @@ def pathCreator(desired_path):
         print(desired_path, " Exists")
 
 def fileChecker(desired_file):
+    global file_exists
     if os.path.isfile(desired_file):
-        print(desired_file + "Exists! moving on...")
+        file_exists=True
     else:
-        print("Downloading " + desired_file)
+        file_exists=False
+    return file_exists
         
 def mp3Download(podcast_dir, podcast_title, episode_link, episode_title, episode_date):
     global file_path
@@ -19,11 +21,14 @@ def mp3Download(podcast_dir, podcast_title, episode_link, episode_title, episode
     download_path = str(podcast_dir + "/" + podcast_title + "/")
     file_path = download_path + episode_date + "-" + episode_title + ".mp3"
     pathCreator(download_path)
-    fileChecker(file_path)
-    if r.status_code == 200:
-        with open(file_path, "wb") as f:
-            f.write(r.content)
+    file_exists = fileChecker(file_path)
+    if file_exists == False:
+        if r.status_code == 200:
+            with open(file_path, "wb") as f:
+                f.write(r.content)
+        else:
+            print(f"Error downloading {file_path}")
     else:
-        print(f"Error downloading {file_path}")
+        print("Episode already downloaded! Moving on...")
 
     return file_path
